Improved the host key trait
This commit changes the return type of the `public_key` method to only
return a slice instead of an owned vector.

This results in fewer allocations needing to be made, but a possibly
longer lived allocation. However for many algorithm implementations, it
may be possible to not require any additional allocations to implement the
`public_key` method, because the keys are already stored in memory.