Remove the abstract `RuntimeState`
It is replaced by fields in the protocol handler itself.
Using a struct just to hold data without making the field public means
very little flexibility when it comes to mutable borrows for the
fields, leading to awkward methods like `output_algorithms_and_rng` or
the `KeyExchangeAccess` struct.

The alternative would have been to keep the `RuntimeState`, but without
methods.
However since the `RuntimeState` was only used in one place (the
`ProtocolHandler`), it makes more sense to instead move its fields
there.